+++ src/main/java/org/jabref/gui/preferences/ai/AiTab.java
@@ -1,6 +1,7 @@
 package org.jabref.gui.preferences.ai;
 
 import javafx.application.Platform;
+import javafx.beans.property.ReadOnlyBooleanProperty;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
@@ -29,6 +30,8 @@ public class AiTab extends AbstractPreferenceTabView<AiTabViewModel> implements 
     private static final String HUGGING_FACE_CHAT_MODEL_PROMPT = "TinyLlama/TinyLlama_v1.1 (or any other model name)";
 
     @FXML private CheckBox enableAi;
+    @FXML private CheckBox autoGenerateEmbeddings;
+    @FXML private CheckBox autoGenerateSummaries;
 
     @FXML private ComboBox<AiProvider> aiProviderComboBox;
     @FXML private ComboBox<String> chatModelComboBox;
@@ -61,6 +64,10 @@ public class AiTab extends AbstractPreferenceTabView<AiTabViewModel> implements 
         this.viewModel = new AiTabViewModel(preferences);
 
         enableAi.selectedProperty().bindBidirectional(viewModel.enableAi());
+        autoGenerateSummaries.selectedProperty().bindBidirectional(viewModel.autoGenerateSummaries());
+        autoGenerateSummaries.disableProperty().bind(viewModel.disableAutoGenerateSummaries());
+        autoGenerateEmbeddings.selectedProperty().bindBidirectional(viewModel.autoGenerateEmbeddings());
+        autoGenerateEmbeddings.disableProperty().bind(viewModel.disableAutoGenerateEmbeddings());
 
         new ViewModelListCellFactory<AiProvider>()
                 .withText(AiProvider::toString)
@@ -186,5 +193,9 @@ public class AiTab extends AbstractPreferenceTabView<AiTabViewModel> implements 
     @FXML
     private void onResetExpertSettingsButtonClick() {
         viewModel.resetExpertSettings();
+    }
+
+    public ReadOnlyBooleanProperty aiEnabledProperty() {
+        return enableAi.selectedProperty();
     }
 }