+++ src/main/java/org/jabref/logic/importer/ImporterPreferences.java
@@ -14,10 +14,10 @@ import javafx.collections.ObservableList;
 import javafx.collections.ObservableSet;
 
 import org.jabref.logic.importer.fileformat.CustomImporter;
+import org.jabref.logic.importer.plaincitation.PlainCitationParserChoice;
 import org.jabref.logic.preferences.FetcherApiKey;
 
 public class ImporterPreferences {
-
     private final BooleanProperty importerEnabled;
     private final BooleanProperty generateNewKeyOnImport;
     private final BooleanProperty warnAboutDuplicatesOnImport;
@@ -26,6 +26,8 @@ public class ImporterPreferences {
     private final ObservableSet<CustomImporter> customImporters;
     private final BooleanProperty persistCustomKeys;
     private final ObservableList<String> catalogs;
+    private final ObjectProperty<PlainCitationParserChoice> defaultPlainCitationParser;
+
     public ImporterPreferences(boolean importerEnabled,
                                boolean generateNewKeyOnImport,
                                Path importWorkingDirectory,
@@ -33,7 +35,9 @@ public class ImporterPreferences {
                                Set<CustomImporter> customImporters,
                                Set<FetcherApiKey> apiKeys,
                                boolean persistCustomKeys,
-                               List<String> catalogs) {
+                               List<String> catalogs,
+                               PlainCitationParserChoice defaultPlainCitationParser
+    ) {
         this.importerEnabled = new SimpleBooleanProperty(importerEnabled);
         this.generateNewKeyOnImport = new SimpleBooleanProperty(generateNewKeyOnImport);
         this.importWorkingDirectory = new SimpleObjectProperty<>(importWorkingDirectory);
@@ -42,6 +46,7 @@ public class ImporterPreferences {
         this.apiKeys = FXCollections.observableSet(apiKeys);
         this.persistCustomKeys = new SimpleBooleanProperty(persistCustomKeys);
         this.catalogs = FXCollections.observableArrayList(catalogs);
+        this.defaultPlainCitationParser = new SimpleObjectProperty<>(defaultPlainCitationParser);
     }
 
     public boolean areImporterEnabled() {
@@ -132,5 +137,17 @@ public class ImporterPreferences {
 
     public ObservableList<String> getCatalogs() {
           return catalogs;
+    }
+
+    public PlainCitationParserChoice getDefaultPlainCitationParser() {
+        return defaultPlainCitationParser.get();
+    }
+
+    public ObjectProperty<PlainCitationParserChoice> defaultPlainCitationParserProperty() {
+        return defaultPlainCitationParser;
+    }
+
+    public void setDefaultPlainCitationParser(PlainCitationParserChoice defaultPlainCitationParser) {
+        this.defaultPlainCitationParser.set(defaultPlainCitationParser);
     }
 }