+++ src/main/java/org/jabref/logic/importer/fileformat/BibliographyFromPdfImporter.java
@@ -39,11 +39,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Parses the references from the "References" section from a PDF
+ * Parses the references from the "References" section from a PDF.
  * <p>
  * Currently, IEEE two column format is supported.
  * <p>
  * To extract a {@link BibEntry} matching the PDF, see {@link PdfContentImporter}.
+ * <p>
+ * TODO: This class is similar to {@link org.jabref.logic.importer.plaincitation.RuleBasedPlainCitationParser}, we need to unify them.
  */
 @AllowedToUseApacheCommonsLang3("Fastest method to count spaces in a string")
 public class BibliographyFromPdfImporter extends Importer {